requirements.txt
python-telegram-bot==20.7
web3==6.11.1
python-dotenv==1.0.0
cryptography==41.0.7
eth-account==0.10.0
eth-typing==4.0.0
eth-utils==2.3.1
requests==2.31.0
aiohttp==3.9.1
python-dateutil==2.8.2
psycopg2-binary==2.9.9 